Add ability to add random jewel crafting recipe.
The jewel crafting table did not have ability to craft and identify random jewels anymore.
This change fixes it. It allows to get a random jewel, if jewel attribute value in config is set to `the_vault:empty`. In that situation it will roll the jewel based on NBT data provided in output NBT.