Optimize Card Tasks Processing in vault.
Original VH implementation processes all cards for all players every tick regardless if player is in the vault or not. This is unnecessary overhead of card deck processing, as well as it suffered from voiding progress, if player overdid it in single tick.

The new implementation starts card tasks only when he enters into vault, and disable card tasks once he exits.
The rewarding is processed only when player exits vault.